feat: security and quality sweep, navigation and performance improvements
Security (CodeQL):
- Antiforgery tokens on the six cookie-authenticated offline/PWA endpoints,
  with the client minting a fresh token per flush. A failed token fetch is now
  retryable rather than a tokenless POST, which the queue classified as a
  permanent 400 and deleted.
- Beta video upload required no authorization at all, and its permission check
  only proved the boulder existed, so any signed-in user could attach a clip to
  any boulder. Authorization now derives from the principal and a server-side
  wall-membership lookup, before the body is read.
- Verification and password-reset codes no longer appear in the mail subject or
  in EmailSender's logs, where an SMTP failure wrote live codes in cleartext.
- Open-redirect targets validated in one place: local paths only, and OAuth
  targets checked against an allow-list built from configuration.

Performance:
- ETag/304 on every image endpoint, with the 304 path resolving metadata via a
  server-side length() projection so it never reads the blob.
- Derived image variants generated on demand and cached on disk, served through
  an allow-listed ?w=; the client picks by CSS size x devicePixelRatio and loads
  the next width up on zoom. Stored originals stay at full resolution -- hold
  detection, alignment and matching read them unchanged.
- Avatars re-encoded to WebP; EF value comparers for the collection properties
  that were being rewritten on every SaveChanges.

Navigation:
- Home always lands on the wall overview, including when already on that wall,
  where the URL does not change and only the carousel needs to move.
- The carousel positions before it is revealed instead of visibly scrolling, and
  repositions on a deep-link change rather than going stale.

Also: correctness and quality fixes across the CodeQL backlog, and the
wall-images volume for gallery uploads and the variant cache.

+++ src/Blocwerk.Core/Services/ProgressionService.cs
@@ -851,7 +851,7 @@ public class ProgressionService : IProgressionService
 
         foreach (var p in pullups)
         {
-            score += p.Sets * p.Repetitions * (1 + (p.AdditionalWeightKg / 10.0));
+            score += (double)p.Sets * p.Repetitions * (1 + (p.AdditionalWeightKg / 10.0));
         }
 
         return score;
